[Redtest] Add a template version for run-redtest
This commit create a cmake template for run-redtest. This template is
named run-redtest.in and allows the future users to easily create their
own.

In order to do so, they have just to copy this template and to
modify the variable defined in the CMakeLists.txt file adjacent to the
template.

Signed-off-by: Armand BENETEAU <armand.beneteau@iot.bzh>